hybrid-search-benchmarks/requirements.txt
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
numpy>=1.21.0
